Resolve #1 and polish code
This change adds support for using the context menu to activate the
extension and with that change it is now also possible to interact with
single links (rather than the whole page).

It also polishes the code a bit and makes the usage of the Chrome sync
storage more obvious.